# openapi.yaml
openapi: 3.0.3
info:
  title: piYès API
  version: 1.0.0
  description: |
    API du service bancaire mobile piYès (Haïti). Base : `/api/v1`.

    Caractéristiques :
    - **Auth** : JWT bearer (24 h) + refresh token httpOnly (30 j). MFA par OTP (5 tentatives, TTL 15 min).
    - **Ledger** : tous les mouvements de fonds passent par un journal en partie double
      (fonctions Supabase `piyes_ledger_post`). Idempotence garantie par la clé
      `Idempotency-Key` (générée automatiquement si absente).
    - **Montants** : en **centimes de gourde HTG** (entiers positifs) dans les requêtes.
    - **Erreurs** : format uniforme `{ "error": { "message", "code" } }`.
    - **Temps réel** : les clients peuvent s'abonner aux changements Supabase
      (User, Transaction, Contact, Notification).

    Vitrine Open Banking : endpoints `/obp/v3.1.0/*` (voir tag `OpenBanking`).
  contact:
    name: piYès
    email: paiements@piyes.ht
servers:
  - url: http://localhost:3000/api/v1
    description: Dev local (backend)
  - url: http://localhost:3000
    description: Dev local (racine – pour OBP)
  - url: https://piyes-backend.vercel.app/api/v1
    description: Production (Vercel)
...
security:
  - bearerAuth: []
paths:
...
  /user/keys:
    post:
      tags: [User]
      summary: Ajouter une clé secondaire (email/téléphone)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Clé ajoutée
...
  /user/keys/{id}:
    delete:
      tags: [User]
      summary: Supprimer une clé secondaire
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Clé supprimée
...
  /obp/v3.1.0/keys:
    post:
      tags: [OpenBanking]
      summary: Créer une clé API tiers (la clé brute n'est affichée qu'une fois)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, description: "Nom de la clé" }
      responses:
        "201":
          description: Clé créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey: { type: string }
                  id: { type: string }
                  name: { type: string }
                  status: { type: string }
    get:
      tags: [OpenBanking]
      summary: Lister les clés API (jamais les hashes)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste des clés
  /obp/v3.1.0/keys/{id}:
    delete:
      tags: [OpenBanking]
      summary: Révoquer une clé API tiers
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Clé révoquée
  /obp/v3.1.0/banks:
    get:
      tags: [OpenBanking]
      summary: Liste des banques (proxy OBP ou mock local)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ apiKeyAuth: [] }]
      responses:
        "200":
          description: Banques
  /obp/v3.1.0/accounts/public:
    get:
      tags: [OpenBanking]
      summary: Comptes publics (proxy OBP ou mock local)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ apiKeyAuth: [] }]
      responses:
        "200":
          description: Comptes publics
  /obp/v3.1.0/banks/{bankId}/accounts/{accountId}/{viewId}/transactions:
    get:
      tags: [OpenBanking]
      summary: Transactions d'un compte (proxy OBP ou mock local)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ apiKeyAuth: [] }]
      parameters:
        - { in: path, name: bankId, required: true, schema: { type: string } }
        - {
            in: path,
            name: accountId,
            required: true,
            schema: { type: string },
          }
        - { in: path, name: viewId, required: true, schema: { type: string } }
      responses:
        "200":
          description: Transactions
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API tiers (Phase 4) — aussi acceptée via Bearer.
...